Change on DDIC check
The DDIC check was limited to objects starting with a 'Z'. This excluded
for example HR objects like PA9005. There are more objects that will not
start with a 'Z', but are still customer developments.
Now the check is done on a higher level. Instead of the object, now the
Development Class that the object belongs to, needs to start with a 'Z'.
This is done by joining the tables DD01L, DD02L and DD04L with table
TADIR.