Add files via upload
Bug fix. Issue: No selections when empty CTSPROJECT.
Trying to prevent this with include/exclude and IN/NOT IN would be annoyingly complex in all occurences. Best option was to re-implemented the previous solution (Add a DUMMY record to the range when the range is empty).

Sometimes, it's better to keep things simple... Thanks @Plabbi for mentioning this bug!